backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-generativeai==0.8.3
python-dotenv==1.0.1
pydantic==2.10.5
pydantic-settings==2.7.1
pytest==8.3.5
